requirements.txt
fastapi==0.115.12
loguru==0.7.3
uvicorn==0.34.0
websockets==15.0.1
pydantic-settings==2.8.1
pydantic==2.11.1
PyJWT==2.10.1